Fix broken imports after moving variableexplorer widgets.
