+++ spyder/plugins/variableexplorer/plugin.py
@@ -14,8 +14,8 @@ from qtpy.QtWidgets import QStackedWidget, QVBoxLayout
 from spyder.config.base import _
 from spyder.api.plugins import SpyderPluginWidget
 from spyder.utils import icon_manager as ima
-from spyder.widgets.variableexplorer.namespacebrowser import NamespaceBrowser
-from spyder.widgets.variableexplorer.utils import REMOTE_SETTINGS
+from spyder.plugins.variableexplorer.widgets.namespacebrowser import NamespaceBrowser
+from spyder.plugins.variableexplorer.widgets.utils import REMOTE_SETTINGS
 from spyder.plugins.variableexplorer.confpage import VariableExplorerConfigPage
