+++ spyder/plugins/variableexplorer/widgets/namespacebrowser.py
@@ -30,10 +30,10 @@ from spyder.utils.misc import fix_reference_name
 from spyder.utils.programs import is_module_installed
 from spyder.utils.qthelpers import (add_actions, create_action,
                                     create_toolbutton)
-from spyder.widgets.variableexplorer.collectionseditor import (
+from spyder.plugins.variableexplorer.widgets.collectionseditor import (
     RemoteCollectionsEditorTableView)
-from spyder.widgets.variableexplorer.importwizard import ImportWizard
-from spyder.widgets.variableexplorer.utils import REMOTE_SETTINGS
+from spyder.plugins.variableexplorer.widgets.importwizard import ImportWizard
+from spyder.plugins.variableexplorer.widgets.utils import REMOTE_SETTINGS
 
 
 SUPPORTED_TYPES = get_supported_types()
