+++ spyder/widgets/externalshell/pythonshell.py
@@ -31,8 +31,8 @@ from spyder.utils.qthelpers import (add_actions, create_action,
                                     create_toolbutton, DialogManager)
 from spyder.widgets.externalshell.baseshell import ExternalShellBase
 from spyder.widgets.shell import PythonShellWidget
-from spyder.widgets.variableexplorer.namespacebrowser import NamespaceBrowser
-from spyder.widgets.variableexplorer.collectionseditor import CollectionsEditor
+from spyder.plugins.variableexplorer.widgets.namespacebrowser import NamespaceBrowser
+from spyder.plugins.variableexplorer.widgets.collectionseditor import CollectionsEditor
 
 
 class ExtPythonShellWidget(PythonShellWidget):
