+++ spyder/widgets/internalshell.py
@@ -33,7 +33,7 @@ from spyder.utils.dochelpers import getargtxt, getdoc, getobjdir, getsource
 from spyder.utils.misc import get_error_match
 from spyder.utils.qthelpers import create_action
 from spyder.widgets.shell import PythonShellWidget
-from spyder.widgets.variableexplorer.objecteditor import oedit
+from spyder.plugins.variableexplorer.widgets.objecteditor import oedit
 # TODO: remove the CONF object and make it work anyway
 # In fact, this 'CONF' object has nothing to do in package spyder/widgets
 # which should not contain anything directly related to Spyder's main app
